Improve label handling for datagrids
This enables datagrid divs to be handled in the same way as containers: if they have a label which is not set to be hidden, then it's rendered as a span, otherwise it is not rendered at all.

n.b. This commit also contains built CSS file from previous PR.